requirements.txt
python-dotenv==0.21.1
requests==2.28.1